* using the ignore_file method to filter out custom action TryTo * avoid calling findCaller twice per log
