+++ screenpy_examples/screenpy_playwright/swaglabs/features/pytest.ini
@@ -6,21 +6,23 @@ python_files = *.py
 ; using *.py  is BROKEN in python 3.10.  
 
 
-addopts = --show-capture=no
-          --capture=no
-          --tb=native
-;          --showlocals
-          --verbosity=0
-          -rA
-          --color=yes
-;          --disable-warnings
+addopts =
+    --show-capture=no
+    --capture=no
+    --tb=native
+;    --showlocals
+    --verbosity=0
+    -rA
+    --color=yes
+;    --disable-warnings
 
 log_cli = on
-log_level = DEBUG
-log_cli_format = %(fileline)-35s %(msg)s
-;log_cli_format = %(asctime)s %(fileline)-25s %(msg)s
-;log_cli_format = %(asctime)s %(levelname)s %(fileline)-25s %(msg)s
-;log_cli_format = %(msg)s
+log_cli_level = INFO
+;log_cli_format = %(message)s
+;log_cli_format = %(name)-8s %(levelname)-5s %(filename)30s:%(lineno)-4s %(message)s
+log_cli_format = %(levelname)-5s %(filename)30s:%(lineno)-4s %(message)s
 
-
-
+# requires ScreenpyLogger
+;log_cli_format = %(name)-8s %(fileline)-35s %(message)s
+;log_cli_format = %(asctime)s %(levelname)-5s %(fileline)-35s %(message)s
+;log_cli_format = %(name)-8s %(levelname)-5s %(fileline)-35s %(message)s
