+++ screenpy_examples/screenpy_selenium/the_internet/features_pytest/pytest.ini
@@ -16,14 +16,14 @@ addopts =
     --color=yes
 ;    --disable-warnings
     -m "not xfail"
-        
 
 log_cli = on
-log_level = DEBUG
-log_cli_format = %(fileline)-35s %(msg)s
-;log_cli_format = %(asctime)s %(fileline)-25s %(msg)s
-;log_cli_format = %(asctime)s %(levelname)s %(fileline)-25s %(msg)s
-;log_cli_format = %(msg)s
+log_cli_level = INFO
+;log_cli_format = %(message)s
+;log_cli_format = %(name)-7s %(levelname)-5s %(filename)35s:%(lineno)-4s %(message)s
+log_cli_format = %(levelname)-5s %(filename)30s:%(lineno)-4s %(message)s
 
-
-
+# requires ScreenpyLogger
+;log_cli_format = %(name)-7s %(fileline)-35s %(message)s
+;log_cli_format = %(asctime)s %(levelname)-5s %(fileline)-35s %(message)s
+;log_cli_format = %(name)-7s %(levelname)-5s %(fileline)-35s %(message)s
